feat(core): market-data decoder and sequence validator (U3)
- decodeFrame(): OKX v5 books-channel frames (snapshot/update action,
  string price/size pairs, ts, optional checksum/seq) and proxy-style
  full-book frames; malformed input yields Malformed status, never
  throws across the API boundary
- SequenceValidator: Sequenced mode enforces strict +1 continuity with
  stale-reject / gap-resync verdicts and counters; Arrival mode stamps
  a monotonic sequence for feeds without one (documented weaker
  guarantee)
- top-of-book CRC32 checksum (IEEE, OKX concatenation rule, depth 25)
  exposed for verification of the applied book state; pinned by the
  standard known-answer vector
- gap_sequence.jsonl fixture triggers exactly one resync signal

+++ core/include/qx/MarketEvent.h
@@ -22,7 +22,7 @@ struct MarketEvent {
     std::vector<Level> bids;
     std::vector<Level> asks;
 
-    std::optional<std::uint32_t> checksum;
+    std::optional<std::int64_t> checksum;
 };
 
 constexpr bool operator==(const MarketEvent& lhs, const MarketEvent& rhs) noexcept
